Call the Gradio backend

Function to call the Gradio backend LLM with the homework_name and student answer.

Usage

get_assignmnent_feedback(homework_name, student_answer, container_name)

Arguments

homework_name

Homework name of format homeworkXqY where X is the homework number and Y is the question number.

student_answer

Student answer.

container_name

Name of container.

Details

The feedback returned is in a funky format because Gradio's API only supports streaming results so we don't get nice clean JSON. Instead we get a line of text with the event status followed by the actual LLM with \n and " escaped something like the following:

"event: complete\ndata: [\"The question is:\\n\\nThe `midwest` data frame...\n]\n\n"

Examples

feedback <- get_assignmnent_feedback(
  homework_name = "homework1-q1",
  student_answer = "give me the question",
  container_name = "container_hostname_as_a_session_id"
)
#> No encoding supplied: defaulting to UTF-8.
feedback
#> [1] "event: complete\ndata: [\"Sure! Here is the question:\\n\\n\\\"The `midwest` data frame that comes with the **ggplot2** package contains demographic characteristics of counties in the Midwest region of the United States. Because the data set is part of the ggplot2 package, you can read documentation for the data set, including variable definitions by typing `?midwest` in the Console or searching for `midwest` in the Help pane.\\n\\nVisualize the distribution of population density of counties in the `midwest` data frame using a histogram with `geom_histogram()` with four separate binwidths: a binwidth of 100, a binwidth of 1,000, a binwidth of 10,000, and a binwidth of 100,000.\\nFor example, you can create the first plot with:\\n\\n```{r}\\n#| label: load-packages\\n#| message: false\\nlibrary(tidyverse)\\n```\\n\\n```{r}\\nggplot(midwest, aes(x = popdensity)) +\\n  geom_histogram(binwidth = 100) +\\n  labs(\\n    x = \\\"Population density\\\",\\n    y = \\\"Count\\\",\\n    title = \\\"Population density of Midwestern counties\\\",\\n    subtitle = \\\"Binwidth = 100\\\"\\n  )\\n```\\n\\nYou will need to make four different histograms.\\nMake sure to set informative titles and axis labels for each of your plots.\\nThen, comment on which binwidth is most appropriate for these data and why.\\\"\"]\n\n"